Move the QueryDescriptions into a separate file
